backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
reportlab
pdfminer.six
sentence-transformers
qdrant-client
pydantic
pymongo
beautifulsoup4
python-dotenv
dnspython
certifi
transformers
torch
numpy
scikit-learn
bson
textblob